+++ CHANGELOG.md
@@ -30,6 +30,9 @@
   strategies for attaching an entity to a repository. The strategies include a `create(...)`
   method that creates entity and attaches entity to the repository and/or an `attach(...)` method
   that manually attaches the entity created outside repository.
+* **\[processor\]** Fix bug in generated components where id was not set if native components are disabled,
+  names are enabled and there is no custom `@ComponentId`. The impact is that the names for these components
+  would all include the name with `0` as id.
 
 ### [v0.88](https://github.com/arez/arez/tree/v0.88) (2018-06-04)
 [Full Changelog](https://github.com/arez/arez/compare/v0.87...v0.88)